+++ src/lazy-file.ts
@@ -1,2 +1,8 @@
 export { type ByteRange, getByteLength, getIndexes } from "./lib/byte-range.js";
-export { type LazyFileContent, LazyFile } from "./lib/lazy-file.js";
+export {
+  type LazyContent,
+  type LazyBlobOptions,
+  LazyBlob,
+  type LazyFileOptions,
+  LazyFile
+} from "./lib/lazy-file.js";